nural/requirements.txt
# 比特币合约量化交易系统依赖库
# Python 3.8+ 推荐使用 Python 3.9 或 3.10

# 核心数据处理
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<3.0.0

# 机器学习（CPU 友好）
lightgbm>=3.3.0,<5.0.0
scikit-learn>=1.0.0,<2.0.0
scikit-optimize>=0.9.0,<1.0.0

# 深度学习（神经网络）
torch>=2.0.0,<3.0.0

# 配置管理
pyyaml>=6.0,<7.0.0

# HTTP 请求（币安 API）
requests>=2.27.0,<3.0.0

# 模型持久化与并行计算
joblib>=1.1.0,<2.0.0

# 数据存储（可选，用于 parquet 格式）
pyarrow>=10.0.0,<15.0.0
# 或者使用 fastparquet 替代
# fastparquet>=0.8.0,<1.0.0
